reworked project to be a toolkit as opposed to a reporting tool
Updated readme to show parsing and db examples.
Added new fields raw and allFields to Parser struct.
Updated tests with more relevant comments.
